feat: Add basic port render scaffold
We're just putting together some JSON that gets sent through a port to be
rendered in JS. We'll have to think about how to model events and after that
we'll look into what we can do to render as efficiently as possible. As the
library doesn't know about control structures we probably have to provide our
own so that we get a full graph of the app that we can then render reactively
only driven by data.